hotel_california/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
django-filter==24.3
sqlparse==0.5.1
